chore: frontend environment variables injection when launching docker inst
As `create-react-app` does not support injecting runtime environment
variables (due to the fact it compiles the code to static html) we are
not able to inject runtinme configs. There are several alternatives
althought the one used here is the one with least amount of codebase
changes. Basically the app is `built` before running. It might make it
slower to start but it will work just fine.